requirements.txt
ecdsa
pycryptodome
galois
numpy
bitstring
python-dateutil